Use uint64_t instead of unsigned.
llvm-svn: 70149

+++ llvm/include/llvm/ADT/PointerIntPair.h
@@ -43,17 +43,17 @@ class PointerIntPair {
   enum {
     /// PointerBitMask - The bits that come from the pointer.
     PointerBitMask =
-      ~(unsigned)(((intptr_t)1 << PtrTraits::NumLowBitsAvailable)-1),
+      ~(uint64_t)(((intptr_t)1 << PtrTraits::NumLowBitsAvailable)-1),
 
     /// IntShift - The number of low bits that we reserve for other uses, and
     /// keep zero.
-    IntShift = (unsigned)PtrTraits::NumLowBitsAvailable-IntBits,
+    IntShift = (uint64_t)PtrTraits::NumLowBitsAvailable-IntBits,
     
     /// IntMask - This is the unshifted mask for valid bits of the int type.
-    IntMask = (unsigned)(((intptr_t)1 << IntBits)-1),
+    IntMask = (uint64_t)(((intptr_t)1 << IntBits)-1),
     
     // ShiftedIntMask - This is the bits for the integer shifted in place.
-    ShiftedIntMask = (unsigned)(IntMask << IntShift)
+    ShiftedIntMask = (uint64_t)(IntMask << IntShift)
   };
 public:
   PointerIntPair() : Value(0) {}
